test(eslint): migrate eslint rule tests from parser to plugin
