+++ packages/eslint-plugin-typescript/tests/lib/eslint-rules/no-dupe-args.js
@@ -1,0 +1,29 @@
+'use strict';
+
+//------------------------------------------------------------------------------
+// Requirements
+//------------------------------------------------------------------------------
+
+const rule = require('eslint/lib/rules/no-dupe-args'),
+  RuleTester = require('eslint').RuleTester;
+
+const ruleTester = new RuleTester({
+  parserOptions: {
+    ecmaVersion: 6,
+    sourceType: 'module',
+    ecmaFeatures: {}
+  },
+  parser: '@typescript-eslint/parser'
+});
+
+ruleTester.run('no-dupe-args', rule, {
+  valid: [
+    // https://github.com/eslint/typescript-eslint-parser/issues/535
+    `
+function foo({ bar }: { bar: string }) {
+    console.log(bar);
+}
+    `
+  ],
+  invalid: []
+});
