+++ packages/eslint-plugin-typescript/tests/lib/eslint-rules/no-redeclare.js
@@ -1,0 +1,34 @@
+'use strict';
+
+//------------------------------------------------------------------------------
+// Requirements
+//------------------------------------------------------------------------------
+
+const rule = require('eslint/lib/rules/no-redeclare'),
+  RuleTester = require('eslint').RuleTester;
+
+const ruleTester = new RuleTester({
+  parserOptions: {
+    ecmaVersion: 6,
+    sourceType: 'module',
+    ecmaFeatures: {}
+  },
+  parser: '@typescript-eslint/parser'
+});
+
+ruleTester.run('no-redeclare', rule, {
+  valid: [
+    // https://github.com/eslint/typescript-eslint-parser/issues/443
+    `
+const Foo = 1;
+type Foo = 1;
+    `,
+    // https://github.com/eslint/typescript-eslint-parser/issues/535
+    `
+function foo({ bar }: { bar: string }) {
+    console.log(bar);
+}
+    `
+  ],
+  invalid: []
+});
