+++ packages/eslint-plugin-typescript/tests/lib/eslint-rules/no-undef.js
@@ -37,7 +37,31 @@ export * from './constants';
 export * from './types';
 
 export default Beemo;
-        `
+    `,
+    // https://github.com/eslint/typescript-eslint-parser/issues/471
+    `
+class X {
+  field = {}
+}
+    `,
+    // https://github.com/eslint/typescript-eslint-parser/issues/466
+    `
+/*globals document, selector */
+const links = document.querySelectorAll( selector ) as NodeListOf<HTMLElement>
+    `,
+    // https://github.com/eslint/typescript-eslint-parser/issues/437
+    `
+interface Runnable {
+  run (): Result
+  toString (): string
+}
+    `,
+    // https://github.com/eslint/typescript-eslint-parser/issues/416
+    `
+export type SomeThing = {
+    id: string;
+}
+    `
   ],
   invalid: []
 });
