+++ packages/typescript-eslint-parser/tests/lib/scope-analysis.ts
@@ -1,4 +1,3 @@
-import { Linter } from 'eslint';
 import fs from 'fs';
 import path from 'path';
 import * as parser from '../../src/parser';
@@ -213,230 +212,5 @@ describe('TypeScript scope analysis', () => {
         expect(scopeTree).toMatchSnapshot();
       });
     }
-  });
-
-  const linter = new Linter();
-  linter.defineParser('@typescript-eslint/parser', parser);
-
-  it('https://github.com/eslint/typescript-eslint-parser/issues/416', () => {
-    const code = `
-export type SomeThing = {
-    id: string;
-}
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-undef': 'error'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue416.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('https://github.com/eslint/typescript-eslint-parser/issues/435', () => {
-    const code = `
-interface Foo {
-    bar: string
-}
-const bar = 'blah'
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-use-before-define': 'error'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue435.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('https://github.com/eslint/typescript-eslint-parser/issues/437', () => {
-    const code = `
-interface Runnable {
-  run (): Result
-  toString (): string
-}
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-undef': 'error'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue437.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('https://github.com/eslint/typescript-eslint-parser/issues/443', () => {
-    const code = `
-const Foo = 1;
-type Foo = 1;
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-redeclare': 'error'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue443.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('https://github.com/eslint/typescript-eslint-parser/issues/459', () => {
-    const code = `
-type foo = any;
-function bar(foo: any) {}
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-shadow': 'error'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('https://github.com/eslint/typescript-eslint-parser/issues/466', () => {
-    const code = `
-/*globals document, selector */
-const links = document.querySelectorAll( selector ) as NodeListOf<HTMLElement>
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-undef': 'error'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('https://github.com/eslint/typescript-eslint-parser/issues/471', () => {
-    const code = `
-class X {
-  field = {}
-}
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-undef': 'error'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('https://github.com/eslint/typescript-eslint-parser/issues/487', () => {
-    const code = `
-export default class Test {
-    private status: string;
-    getStatus() {
-        return this.status;
-    }
-}
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-restricted-globals': ['error', 'status']
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('1: https://github.com/eslint/typescript-eslint-parser/issues/535', () => {
-    const code = `
-function foo({ bar }: { bar: string }) {
-    console.log(bar);
-}
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-dupe-args': 'error',
-        'no-redeclare': 'error'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('2: https://github.com/eslint/typescript-eslint-parser/issues/535', () => {
-    const code = `
-import {
-  observable,
-} from 'mobx';
-
-export default class ListModalStore {
-  @observable
-  orderList: IObservableArray<BizPurchaseOrderTO> = observable([]);
-}
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-unused-vars': 'error'
-      },
-      parserOptions: {
-        sourceType: 'module'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('https://github.com/eslint/typescript-eslint-parser/issues/550', () => {
-    const code = `
-function test(file: Blob) {
-  const slice: typeof file.slice =
-    file.slice || (file as any).webkitSlice || (file as any).mozSlice
-  return slice
-}
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      rules: {
-        'no-use-before-define': 'error'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue.ts' });
-
-    expect(messages).toStrictEqual([]);
-  });
-
-  it('https://github.com/bradzacher/eslint-plugin-typescript/issues/255', () => {
-    const code = `
-window.whatevs = {
-  myFunc() {
-    console.log('yep');
-  }
-};
-`;
-    const config: Linter.Config = {
-      parser: '@typescript-eslint/parser',
-      parserOptions: {
-        sourceType: 'module'
-      },
-      rules: {
-        strict: 'error'
-      }
-    };
-    const messages = linter.verify(code, config, { filename: 'issue255.ts' });
-
-    expect(messages).toStrictEqual([]);
   });
 });
